'Adds database connection and nodemon dev script'
This commit includes the addition of a database connection using Mongoose to connect to a MongoDB cluster. The connection details are specified in the `connection.js` file. The commit also adds a `dev` script to the `package.json` file, which can be used to run the server with nodemon for development.